add CalendarController tests
split rest api get calendar into calendar (by year) and calendars
replace int year with Short year
hashmap<year, GrandPrixModel> instead of list of GrandPrixModel in CalendarModel
remove JsonCreator constructors

+++ src/main/java/sk/f1api/f1api/entity/Season.java
@@ -8,7 +8,6 @@ import jakarta.persistence.*;
 import jakarta.persistence.criteria.CriteriaBuilder;
 import jakarta.persistence.criteria.CriteriaQuery;
 import jakarta.persistence.criteria.Join;
-import jakarta.persistence.criteria.JoinType;
 import jakarta.persistence.criteria.Predicate;
 import jakarta.persistence.criteria.Root;
 import lombok.Getter;
@@ -84,7 +83,7 @@ public class Season implements Identifiable {
 				criteriaComponents.getRoot(), currentVersionId);
 	}
 
-	public static List<Season> loadByYear(Session session, Integer currentVersionId, int year) {
+	public static List<Season> loadByYear(Session session, Integer currentVersionId, Short year) {
 		CriteriaComponents<Season> criteriaComponents = CriteriaUtil.getCriteriaComponents(session, Season.class);
 
 		return load(
